Model

BlockOpt.Model — Type

Model

Specifies the unconstrianed minimization of a smooth objective function. A model is minimally constructed with a name and may be incrementally loaded. Once a model is loaded, the objective, and gradient function may no longer be modified. A model creates a directory, which stores logged information throughout an instances life.

Interface

BlockOpt.name — Function

name(m::Model)

The name of model m.

BlockOpt.objective — Function

objective(m::Model)

The objective function of model m, having an unassigned default value of missing.

BlockOpt.gradient — Function

gradient(m::Model)

The gradient function of model m, having an unassigned default value of missing.

BlockOpt.initial_iterate — Function

initial_iterate(m::Model)

The initial iterate of model m, having an unassigned default value of missing.

BlockOpt.formula — Function

formula(m::Model)

The formula of model m, having an unassigned default value of missing.

BlockOpt.dimension — Function

dimension(m::Model)

The dimension of model m, which is missing so long as the initial iterate isn't specified.

BlockOpt.directory — Function

directory(m::Model)

The directory path of model m, with the relative portion given by the name of m.

BlockOpt.objective! — Function

objective!(m::Model, f)

Assign model m the objective function f.

If the model is final, the call simply returns without modifying the model.

BlockOpt.gradient! — Function

gradient!(m::Model, ∇f!)

Assign model m the in-place gradient function ∇f!, of the form

∇f!(out, x) = (out .= ∇f(x))

where ∇f(x) is the steepest descent direction at x in stored the place of the inputed buffer out.

If the model is final, the call simply returns without modifying the model.

BlockOpt.initial_iterate! — Function

initial_iterate!(m::Model, x0)

Assign model m an initial starting location, ideally a resonable guess of a minima for the models objective function.

BlockOpt.formula! — Function

formula!(m::Model, f)

Assign model m an escaped $\LaTex$ string, for example

dot_formila = "$ f(x) = x⋅x $" 

formula!(out, dot_formila)

assigns the euclidean squared distance as the models objective function formula.

BlockOpt.obj — Function

obj(m::Model, x)

Evaluates the objective function of model m at x.

BlockOpt.grad! — Function

grad!(m::Model, out, x)

Evaluates the in-place gradient function of model m at x, storing the steepest descent direction in the place of input buffer out.

BlockOpt.hessAD — Function

hessAD(m::Model, x)

The dense Hessian matrix of model m's objective function at the point x. The computation uses ForwardDiff.jacobian forward-mode AD function on the model's gradient function.

BlockOpt.hess_sample — Function

hess_sample(m::Model, x, dx)

The Hessian vector-product of model m's objective function at the point x with the vector dx.