Options

BlockOpt.DriverOptions — Type

DriverOptions

Specifies Driving parameters used in a Simulation instance.

Interface

The DriverOptrions type is best thought of as the mutable protion of a Driver instance. The methods below are simple fallbacks of the forwarded Driver methods.

Note

It is possible to pass a DriverOptions instance as the options keyword argument to a Driver constructor.

BlockOpt.samples — Method

samples(o::DriverOptions)

The number of hessian samples taken at each succussful step in a simulation.

BlockOpt.Δ_max — Method

Δ_max(o::DriverOptions)

The maximum trust-region radius of a driven simulation.

BlockOpt.δ_tol — Method

δ_tol(o::DriverOptions)

The relative tolerance passed to pinv while performing a block QN update.

BlockOpt.ϵ_tol — Method

ϵ_tol(o::DriverOptions)

The absolute convergence tolerance, occuring at $xₖ$ such that $||∇f(xₖ)|| ≤ ϵ$.

BlockOpt.max_iterations — Method

max_iterations(o::DriverOptions)

The maximum number of iterations for a driven simulation.

BlockOpt.weave_level — Method

weave_level(o::DriverOptions)

Returns the current weave level of a simulation.

BlockOpt.log_level — Method

log_level(o::DriverOptions)

Returns the current logging level of a simulation

BlockOpt.samples! — Function

samples!(o::DriverOptions, s)

Set the number of hessian samples collected during each succussful step to some even natrual s where $s = 2w$.

BlockOpt.Δ_max! — Function

δ_tol!(o::DriverOptions, δ)

Set the maximum trust-region radius to some positive Δ.

BlockOpt.δ_tol! — Function

δ_tol!(o::DriverOptions, δ)

Set the pinv relative tolerance used in the QN update to some positive δ.

BlockOpt.ϵ_tol! — Function

ϵ_tol!(o::DriverOptions, ϵ)

Set the terminal convergence tolerance to some positive ϵ.

BlockOpt.max_iterations! — Function

max_iterations!(o::DriverOptions, K)

Set the terminal iteration to the positive integer K.

BlockOpt.weave_level! — Function

weave_level!(o::DriverOptions, level::WeaveLevel)

Set the weave level to NONE or ALL, toggiling the optional Weave.jl generated report of a simulation.

BlockOpt.log_level! — Function

log_level!(o::DriverOptions, level::LogLevel)

Assign the simulation logging level to INFO, DEBUG, WARN, or ERROR.