Driver

BlockOpt.Driver — Type

Driver

Specifies the driving parameters of a Simulation instance. A driver is assigned an immutable S_update, QN_update, and pflag upon construction.

Interface

BlockOpt.SR1 — Function

SR1

Returns the algebraically mininimal SR1 inverse Quasi-Newton block update satisfying the inverse multi-secant condition $H ⋅ V = U$, where δ is the Moore-Penrose psuedoinverse relative tolerance.

See: Algorithm $4.2.$

BlockOpt.PSB — Function

PSB

Powell-Symmetric-Broyden generalized Quasi-Newton block update, where δ is the Moore-Penrose psuedoinverse relative tolerance.

See: Algorithm $4.3.$

BlockOpt.S_update_a — Function

S_update_a

Random set of orthonormal sample directions.

See: Equation $(6.1.a)$.

BlockOpt.S_update_b — Function

S_update_b

Random set of sample directions orthogonal to the pervious sample space given by input Sₖ.

See: Equation $(6.1.b)$.

BlockOpt.S_update_c — Function

S_update_c

Attempts to guide algorithm to accurately resolve eigen-space associated with the larger Hessian eigenvalues.

See: Equation $(6.1.c)$.

BlockOpt.S_update_d — Function

S_update_d

Variant of $(6.1a)$ that includes approximate curvature information along the previously choosen step.

See: Equation $(6.1.d)$.

BlockOpt.S_update_e — Function

S_update_e

Variant of $(6.1b)$ that includes approximate curvature information along the previously choosen step.

See: Equation $(6.1.e)$.

BlockOpt.S_update_f — Function

S_update_f

Variant of $(6.1c)$ that includes approximate curvature information along the previously choosen step.

See: Equation $(6.1.f)$.

BlockOpt.pflag — Function

pflag(d::Driver)

The preliminary secant QN update flag of driver D.

See: SR1, PSB

BlockOpt.QN_update — Function

QN_update(d::Driver)

The QN update formula of Driver d.

See: SR1, PSB

BlockOpt.S_update — Function

S_update(d::Driver)

The supplemental sample direction update formula of Driver d.

See: S_update, S_update_a, S_update_b, S_update_c, S_update_d, S_update_e, S_update_f