Model Interface

BlockOpt.name — Function

name(m::Model)

The name of model m.

BlockOpt.objective — Function

objective(m::Model)

The objective function of model m, having an unassigned default value of missing.

BlockOpt.gradient — Function

gradient(m::Model)

The gradient function of model m, having an unassigned default value of missing.

BlockOpt.initial_iterate — Function

initial_iterate(m::Model)

The initial iterate of model m, having an unassigned default value of missing.

BlockOpt.formula — Function

formula(m::Model)

The formula of model m, having an unassigned default value of missing.

BlockOpt.dimension — Function

dimension(m::Model)

The dimension of model m, which is missing so long as the initial iterate isn't specified.

BlockOpt.directory — Function

directory(m::Model)

The directory path of model m, with the relative portion given by the name of m.