Options

BlockOpt.DriverOptions — Type

DriverOptions

Specifies simulation options.

Interface

The DriverOptrions type is best thought of as the mutable protion of a Driver instance. The methods below are forwarded to the Driver, which has a DriverOptions instance.

Note

It is possible to pass a DriverOptions instance as the options keyword argument to the Driver constructor.

This is useful when constructing mutiple drivers with the same set of options.

BlockOpt.samples — Method

samples(o::DriverOptions)

The number of hessian samples taken at each successful step in a simulation. The returned value corresponds to $2w$.

BlockOpt.Δ_max — Method

Δ_max(o::DriverOptions)

The maximum trust-region radius allowed in simulation.

BlockOpt.δ_tol — Method

δ_tol(o::DriverOptions)

The relative tolerance passed to pinv while performing a block QN update.

BlockOpt.ϵ_tol — Method

ϵ_tol(o::DriverOptions)

The absolute convergence tolerance of a successful simulation, occuring at iterate $xₖ$ when $||∇f(xₖ)|| ≤ ϵ$.

BlockOpt.max_iterations — Method

max_iterations(o::DriverOptions)

The maximum number of iterations for a driven simulation.

BlockOpt.samples! — Function

samples!(o::DriverOptions, s)

Set the number of hessian samples collected during each successful step to some even natural number. The set value corresponds to $2w$ in the article.

BlockOpt.Δ_max! — Function

δ_tol!(o::DriverOptions, δ)

Set the maximum trust-region radius to some positive Δ.

BlockOpt.δ_tol! — Function

δ_tol!(o::DriverOptions, δ)

Set the pinv relative tolerance used in the QN update to some positive δ.

BlockOpt.ϵ_tol! — Function

ϵ_tol!(o::DriverOptions, ϵ)

Set the terminal convergence tolerance to some positive ϵ.

BlockOpt.max_iterations! — Function

max_iterations!(o::DriverOptions, K)

Set the terminal iteration to the positive integer K.