Simulation

BlockOpt.Simulation — Type

Simulation

Enty point and exit point of an iteration. A simulation composes a iterations trace and backend iteration of Algorithm 7.1. An instance of a simulation is returned from an optimize call.

Interface

BlockOpt.trs_timer — Function

trs_timer

The elapsed time simulation s has spent in trs_solve(s).

BlockOpt.trs_counter — Function

trs_counter

The count of trust region subproblem subsolves of simulation s.

BlockOpt.ghs_timer — Function

ghs_timer

The elapsed time simulation s has spent in gHS(s).

BlockOpt.ghs_counter — Function

ghs_counter

The number of gHS evaluations of simulation s.

BlockOpt.f_vals — Function

f_vals

A vector storing objective values $f(xₖ)$ for each iterate $xₖ$.

f_vals gets updated at each accepted step.

BlockOpt.∇f_norms — Function

∇f_norms

A vector storing normed gradient values $||∇f(xₖ)||₂$ for each iterate $xₖ$.

∇f_norms gets updated at each accepted step.

BlockOpt.p_norms — Function

p_norms

A vector storing the distance of each step $||pₖ||₂$.

p_norms gets updated at each accepted step.

BlockOpt.Δ_vals — Function

Δ_vals

A vector storing the trust-region radius passed to trs_small of TRS.jl, during each succussful trust-region subproblem solve.

Δ_vals gets updated at each accepted step.

BlockOpt.ρ_vals — Function

ρ_vals

A vector storing the ratio of actual reduction to model reduction of each successful step.

ρ_vals gets updated at each accepted step.